tailscale.nuspec, tools: remove vestigial content and modernize package scripts
Much of the content in our package scripts was for dealing with NSIS and our
own WinTun, neither of which exist anymore in modern versions of Tailscale.

We make a few minor updates to tailscale.nuspec to keep its contents up
to date.

We remove tailscale.cer and the certificate installation script, as those
are no longer necessary.

We remove the uninstall script, as we may simply rely on Chocolatey's
AutoUninstall functionality (which groks MSI).

We update the beforemodify script to remove any previously installed
certificate. We remove the old bits that would explicitly stop tailscaled,
as this is always done in the MSI.

Tested install, uninstall, upgrade from MSI, upgrade from NSIS (1.34.0).

Fixes https://github.com/tailscale/tailscale/issues/10839

Signed-off-by: Aaron Klotz <aaron@tailscale.com>